Separate main file to multiple source files.
